+++ src/main/java/com/github/ajouteam6/premierleague/domain/team/TeamStat.java
@@ -1,4 +1,4 @@
-package com.github.ajouteam6.premierleague.domain;
+package com.github.ajouteam6.premierleague.domain.team;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
